Removed backup file + Removed debuggin prints
