+++ server/services/parcel_finder_service.py
@@ -1,5 +1,5 @@
 from sigpac_tools.find import find_from_cadastral_registry
-from ..utils.parcel_finder_utils1 import *
+from ..utils.parcel_finder_utils import *
 import os
 
 def get_parcel_image( cadastral_reference: str, date: str) -> tuple:
@@ -19,38 +19,23 @@ def get_parcel_image( cadastral_reference: str, date: str) -> tuple:
     # Get parcel data
     geometry, metadata = find_from_cadastral_registry(cadastral_reference)
     
-    print("PARCEL DATA FOUND!") if geometry else None
-
     # Get GeoJSON data and dataframe and list of UTM zones
     geojson_data, gdf = get_geojson_data(geometry, metadata)
     zones_utm = get_tiles_polygons(gdf)
     list_zones_utm = list(zones_utm)
 
-    print("ZONES FOUND!") if geometry else None
-
     # Download RGB image:
     rgb_images_path = download_tiles_rgb_bands(list_zones_utm, year, month)
     
-    print("RGB BANDS FOUND!", rgb_images_path) if len(rgb_images_path) > 1 else print("NO RGB BANDS FOUND!")
-
     if not rgb_images_path:
         print("No images are available for the selected date, images are processed at the end of each month.")
         return None, None
 
-
     out_dir, png_paths, rgb_tif_paths = get_rgb_parcel_image(cadastral_reference, geojson_data, rgb_images_path)
-
-    print("RGB BANDS MERGED!") if rgb_images_path else None
-
-    print("out_dir", out_dir) if out_dir else None
-    print("png_paths", png_paths) if png_paths else None
-    print("rgb_tif_paths", rgb_tif_paths) if rgb_tif_paths else None
 
     #TODO: Get image from geometry (image-workflow.pptx)
     
     sigpac_image_name = png_paths.pop()  # there should only be one file
-
-    print("SIGPAC IMAGE NAME!", sigpac_image_name) if sigpac_image_name else None
 
     sigpac_image_url = f"{os.getenv('API_URL')}/uploads/{os.path.basename(sigpac_image_name)}"
 
@@ -74,9 +59,6 @@ def get_rgb_parcel_image(cadastral_reference, geojson_data, rgb_images_path_valu
     Note:
         This function relies on external functions `cut_from_geometry` and `rgb` to perform cropping and image processing.
     """
-    print("RGB IMAGES PATH:", rgb_images_path_values) if rgb_images_path_values else None
-    print("GEOJSON DATA RECIEVED:", geojson_data.keys()) if geojson_data else None
-
     unique_formats = list(
         set(
             f.split(".")[-1].lower()
@@ -84,14 +66,11 @@ def get_rgb_parcel_image(cadastral_reference, geojson_data, rgb_images_path_valu
             if isinstance(f, str) and "." in f
         )
     )
-    print("UNIQUE FORMATS:", unique_formats) if unique_formats else None
-    
     if len(unique_formats) > 1:
         raise ValueError(
             f"Unsupported format. You must upload images in one unique format."
         )
-    print("UNIQUE VALUE:", unique_formats) if unique_formats else None
-
+    
     # Crop the parcel outline using the geomatry available
     cropped_parcel_masks_paths = []
     for feature in geojson_data["features"]:
